# procstat speed improvement