darwin:
before
```
$ go test -short ./plugins/inputs/procstat/... -bench=. -benchmem
goos: darwin
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/inputs/procstat
BenchmarkPattern-12        	       1	1600582288 ns/op	 5874216 B/op	   48391 allocs/op
BenchmarkFullPattern-12    	       1	3620761444 ns/op	11612472 B/op	   86876 allocs/op
PASS
ok  	github.com/influxdata/telegraf/plugins/inputs/procstat	5.450s
```
after
```
$ go test -short ./plugins/inputs/procstat/... -bench=. -benchmem
goos: darwin
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/inputs/procstat
BenchmarkPattern-12        	      57	  25058656 ns/op	  309886 B/op	    2991 allocs/op
BenchmarkFullPattern-12    	       1	1798469106 ns/op	 6176720 B/op	   41895 allocs/op
PASS
ok  	github.com/influxdata/telegraf/plugins/inputs/procstat	3.466s
```