@danielnelson 2020-06-16
Linux comparisons

```
go test -short ./plugins/inputs/procstat --bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/inputs/procstat
BenchmarkPattern-4                    14          80332678 ns/op        20518161 B/op     215585 allocs/op
BenchmarkFullPattern-4                14          80817491 ns/op        20781215 B/op     216293 allocs/op
PASS
ok      github.com/influxdata/telegraf/plugins/inputs/procstat  2.450s
```

```
go test -short ./plugins/inputs/procstat --bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/inputs/procstat
BenchmarkPattern-4                  2005            601808 ns/op          105720 B/op       1636 allocs/op
BenchmarkFullPattern-4               655           1869698 ns/op          369833 B/op       2349 allocs/op
PASS
ok      github.com/influxdata/telegraf/plugins/inputs/procstat  2.712s
```